services/inference/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.32
pydantic>=2.8,<3
python-multipart>=0.0.9,<0.1
python-dotenv>=1.0,<2
Pillow>=10,<12
numpy>=2.0,<3
opencv-python-headless>=4.10,<5
openpyxl>=3.1,<4
rapidocr-onnxruntime>=1.4,<2
easyocr>=1.7,<2
openai>=1.0,<3
pytest>=8.3,<9
